feat: add show countries tests
This commit introduces a new test file, `show_countries_test.dart`, which contains widget tests for the `ShowCountries` functionality. The tests cover the following scenarios:

- Displaying the modal bottom sheet with a list of countries.
- Customizing the displayed information (flag, name, dial code).
- Selecting a country and returning the correct `Country` object.
- Applying a custom shape to the `ListTile` widgets.

These tests ensure the correct rendering and behavior of the `ShowCountries` widget. A skip has been added to the test that checks the selected country because the equals method is not implemented in the Country model.