feat: add equatable dependency
This commit adds the `equatable` dependency to the `pubspec.yaml` file.

The `equatable` package simplifies value equality in Dart. It will be used to easily compare instances of data classes, specifically the `Country` class, by simply listing the properties that make up the object's identity. This avoids having to manually override the `==` operator and `hashCode` getter.